增加linux构建, 测试通过，并修复部分编译错误.

+++ DBAccess/mysql_api/MysqlConnection.h
@@ -37,7 +37,7 @@ public:
 	// 取得上一步 INSERT 操作产生的 ID
 	// oracle 采用sequence（序列）来获取最后插入行ID，需要指定序列名
 	// mysql 序列名无效
-	virtual bool GetLastInsertID( const char* szSeqName, signed __int64& lRowID ) override;
+	virtual bool GetLastInsertID( const char* szSeqName, signed long long& lRowID ) override;
 
 	// 执行查询sql,返回数据集
 	virtual	IRecordSet* ExecuteQuery( const char* szSql ) override;
